Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -4,8 +4,8 @@ int main()
 {
     float a,b;
     scanf("%f%f",&a,&b);
-    int harshit = a * b;
-    printf("Product: %.2d",harshit);
+    float harshit = a * b;
+    printf("Product: %.2f",harshit);
 
     return 0;
